task(refactor and cleanup): Move reusable pieces out and refactor Move the factory generated from callbackMap to a new file to clean up the component file
Fix some set states so that batching is better updated with React
